Add Reply Bot and Service Bot
Bots need to have consumer and app credentials added before running, in addition to the Twitter ID of the bot account. Right now, they are configured to run just once and process their home page in a batch, but it can be set up to run "continuously" by changing the range on line 88 to an arbitrarily high number from just once. Tweets that have already been answered will not be answered again.